New converter: memprof2calltree
Memprof is a nice Memory Usage Profiler
(http://www.gnome.org/projects/memprof). It has a GUI, but
the data visualization is (of course) not as good as KCachegrind's.
You can create snapshots of memory usage of unmodified binaries
similar to Valgrind, save the profile, convert it and load it
into KCachegrind. (Convertion of Leak Info is currently not supported)

svn path=/trunk/kdesdk/kcachegrind/; revision=310096