KCachegrind: Fix slow loading times of unusual profile data files
Running callgrind on a Valgrind tool (this is valgrind hosting itself)
produces such unusual profiles, where the number of callers or callees
of functions can reach some 100 thousands (or even more).

This fix gets rid of load time complexity quadratic to the number of
callers/callees of a function, which made KCachegrind unusable on such
profiles before.


svn path=/branches/KDE/3.5/kdesdk/kcachegrind/; revision=536128