kcachegrind: fix hex code column with in InstrView
When hex code mode was disabled on filling the InstrView
with assembly, we did not get any width of a hex column,
because it was set manually to width zero.
Enabling the hex code mode afterwards did *not* show the
column, because we still set the column width to zero.

Now, on refresh of InstrView data, we set the mode of the
hex code column to automatic resizing, get the width, and
reset to 0 if hex mode is off.

svn path=/trunk/KDE/kdesdk/kcachegrind/; revision=694392