kcachegrind: emit loading errors as signals
Use more flexible interface in loaders for profile data
by using signals instead of directly using kDebug().
Current functionality is not changed, as the given loading
errors still are output via kDebug(), but now from TopLevel
class. But this allows to use a dialog or dock later for
showing loading errors instead of on console.

svn path=/trunk/KDE/kdesdk/kcachegrind/; revision=712594