kcachegrind: Fix automatic position of birds-eye view of call graph
This partly was screwed up because the qt3to4 script replaced
e.g. TopLeft of a enum in CallGraphView to "Qt::TopLeftCorner",
which is way off.

This also adds a further "position", namely to hide to birds-eye view.

And finally, this gets rid of one use of Q3PopupMenu by making
QAction's from the menu items to choose the position of the birds-eye
view.

svn path=/trunk/KDE/kdesdk/kcachegrind/; revision=716508