kcachegrind: port context menu of call graph to QMenu (part 1)
This convertion to QMenu/QAction also gets rid of some
sanity checking of configuration parameters, which
(1) was misplaced here and (2) is not needed.

Because of the huge context menu of the call graph, this is done
in a few steps, ie. further parts coming...

I admit that the resulting code is probably easier to
understand (no fiddling with menu item IDs), and no large
switch statement any more.
But the introduction of a lot of new QActions and new slots
to handle triggered actions makes the code not really smaller.

svn path=/trunk/KDE/kdesdk/kcachegrind/; revision=716970