kcachegrind: Port context menu of call graph to QMenu/QAction (part 2)
This fixes a bug with the "Minimum Call Cost". Previously, it was
an absolute value as the "Minimum Node Cost". But we can not
preserve the call/node cost ratio this way when switching to an unlimit
node cost and back to a limited one. The result was the the call limit
was unlimited afterwards.
This is fixed by storing the "Minimum Call Cost" directly as ratio.

svn path=/trunk/KDE/kdesdk/kcachegrind/; revision=716971