kcachegrind: cleanup kcachegrind specific icons
The *rec.png icons were not used since a long time,
so remove them. But the icon for "relative to parent"
is not provided in the default icon set, so add it
(while the "move" action icon for this setting is
a little obscure, it was used in kcachegrind from
the beginning, so use it for KDE4, too).

svn path=/trunk/KDE/kdesdk/kcachegrind/; revision=753639