KCachegrind: restore correct field drawing in RectDrawing
I still do not have a good solution for bug 161276.
If I change the drawing order to first draw the percentage
into call graph nodes, they always will be shown. But then,
in the compact mode, the symbols are not shown, which IMHO
is worse than not showing the percentage.
As this all very much depends on the size/type of the system
font, the best solution probably is to be able to freely
resize the node rectangles...

svn path=/trunk/KDE/kdesdk/kcachegrind/; revision=814592