KCachegrind: fix Q3PtrList -> QList ports
The loop form for Qt3 lists using "for(i=l.first();i;i=i->next())"
can not be ported to Qt4 lists with "foreach(i, l)" when the
item <i> is used after the loop. This introduced 3 bugs in recent
commits! Better always embed the type in "foreach(type i, l)".
This way, we can be sure that <i> is not used afterwards.

svn path=/trunk/KDE/kdesdk/kcachegrind/; revision=1001700