Add color settings page to qcachegrind configuration.
Now KCachegrind and QCachegrind configuration dialogs provide
the same configuration settings. However, QCachegrinds dialog
allows to jump to specific config items from the outside.

Therefore, the next step is to use QCachegrinds config dialog
also in KCachegrind. However, as this will add/remove a bunch
of strings, this has to wait for KDE 4.5.

svn path=/trunk/KDE/kdesdk/kcachegrind/; revision=1054311