Add global option to hide C++ template parameters in symbols
This adds a toolbar button (+icon) and a menu item under "View".
When template parameters are hidden, you get the full symbol
by hovering over the active symbol name label. The later even
shows the structure of a C++ symbol.

Works both with KCachegrind and QCachegrind.

(done with the help of / partly by Michael Dengler)

svn path=/trunk/KDE/kdesdk/kcachegrind/; revision=1078313