Fix inconsistant state after resetting search
With a symbol filter, the flat profile only shows a subset
of functions matching the search filter. This filter should
be reset when a function is activated from outside of the
flat profile (ie. in visualization view).

Previously, only the search test was cleared, leading to
an inconsistent state, as the previous filter was still active,
and a newly activated function was not selected if not in the
filtered subset.

svn path=/trunk/KDE/kdesdk/kcachegrind/; revision=1135242