qcachegrind.pro: Fix for own moc generator
To allow the qmake-only qcachegrind target to share code with
the cmake-driven kcachegrind target, we need to use the same
moc file naming conventions as KDE. For that, we define our
own new_moc compiler, generating moc files with KDE name convention.
The depend_command used to run gcc/sed, but that never worked, as
we do not have the include paths available yet. Only with Qt 4.7,
the errors generated by depend_command are forwarded to the terminal,
and were surpressed before. So this bug never was visible.

The solution is to just print out the file name containing the
QObject subclass definition, as moc output only depends on this (?).
This gets rid of the scary amount of errors running qmake from 4.7.

svn path=/trunk/KDE/kdesdk/kcachegrind/; revision=1238680