Fix dump requests for VG > 3.7
KCachegrind always supported triggering a new dump of a
Callgrind run in the background. The manual way to do this
was to write a command file, which regularily was polled
for existance in Callgrind.
Up to VG 3.6.x, callgrind_control itself did exactly this.
VG 3.7.0 introduced a new way for interactive control of
Valgrind tools via "vgdb", which uses ptrace, and
callgrind_control uses this mechanism now. With VG 3.8,
Callgrind will remove support for polling a command file.
To make KCachegrind still work, it now uses callgrind_control.

svn path=/trunk/KDE/kdesdk/kcachegrind/; revision=1264196