Allow instructions/source lines to be activated
The selection mechanism after activation was buggy.
Now we just reset the selection after activation, and
every view may decide on its own about a useful selection
on updating its contents.

The term "activation" is used in KCachegrind when e.g. a new
function is shown in all visualization views; a "selection" is
just a selected item within a visualization.