Better support for ARM
When calling objdump on an file to get machine code,
(1) use $OBJDUMP as command instead of "objdump" if set,
(2) prepend $SYSROOT to an absolute file name if set.

Further, if "arch: arm" is found in header of a profile dump,
set bit0 of all addresses in InstrView to 0. Bit0 is used in
Arm32 to specify encoding, but instructions always start at
even addresses. If bit0 is 1, this is Thumb encoding, but
objdump detects this automatically and wants real addresses of
instructions.