Make source copy/paste'able with double-click
It may be useful to copy some source from SourceView with the mouse.
This change makes a line change to read-only QLineEdit for selection
when double-clicking.

To be able to also copy source file name, make it its own item,
instead of displaying the file name in the header field.