Fix detection of callgrind files for very long commands
Backport fix from master

A file created by a callgrind run of "ls /usr/bin/*" was not detected
as callgrind file due to the previous heuristics that the first 2k
characters should contain "events:".
Also checking for "creator: callgrind" fixes this issue.

This assumes that other creators of callgrind files never have such long
command lines, are put the "event:" line before the "cmd:" line.

The real fix would be a mandatory first magic line, which unfortunately is
not part of the format specification, and now we need to be compatible.