Correctly load the Qt-native translations
Summary:
ecm_create_qm_loader() uses Q_COREAPP_STARTUP_FUNCTION to inject the
loader of the additiona translator. The macro does not seem to work
when the loader is linked in a separate static library (the code
which uses Qt translations is all inside libcore and libviews).
So call the ECM macro (and then link the special loader) where
the Q(Core)Application is defined.
Many previously untranslated strings are not translated.

Test Plan:
The code compiles, and content of the right pane (coming from
libviews and libcore) is now translated.

Reviewers: #localization, weidendo

Reviewed By: weidendo

Subscribers: weidendo, aacid, #localization

Differential Revision: https://phabricator.kde.org/D4795