Use a common save dialog for exporting call graphs
This lets you pick where the saved call graph visualization should
go, and what format you wanted it in. Previously, it generated a
dot file, and on Unix, a PostScript file. This lets you choose a
dot or PS/PDF file, rendering to an intermediate dot file if needed.

This also changes the context menu in the per-function graph too,
increasing commonality in code.

More formats can be added as needed.

Possible caveats:
- QProcess might be blocking.