Open folder browser on adding first source annotation item
User testing shown that this wasn't obvious. When the user adds the
first item (usually for all objects), open the dialog to make it
obvious, otherwise they tend to stumble around looking for more
options. For more items, it won't do this because it's assumed
that it's for different objects. (My testing with users didn't involve
programs with object files though.)

Can be replicated with KCG.