Add zooming in and out in call graph view
This change allows the user to use 'ctr + mouse-wheel' to zoom in and
out the call graph view. It also sets the maximum and minimum zoom
level.

I find that the default scale for the call graph view has always been
a little bit too large which results in a lot of panning around to see
what the flow of call stack.