Quote the argument to the filename in instrview call to objdump
This allows for space characters in the path.
A more robust way to do this would be passing in a string list like ObjdumpParser::runObjdump in controlflowgraphview.cpp does it - as it is my fix will still fail on paths with double-quotes - but instrview uses a deliberately customizable template string, see commit 1a504f73921164a85b0949b81cb3a76d75e3aa05.